// types.ts
export interface RPMFormData {
  namaSatuanPendidikan: string;
  namaGuru: string;
  nipGuru: string;
  namaKepalaSekolah: string;
  nipKepalaSekolah: string;
  jenjangPendidikan: string;
  kelas: string;
  mataPelajaran: string;
  capaianPembelajaran: string;
  materiPelajaran: string;
  jumlahPertemuan: number;
  durasiSetiapPertemuan: string;
  praktikPedagogis: string[];
  dimensiLulusan: string[];
}